pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devops-agent"
version = "0.1.0"
description = "Telegram-controlled, LangGraph-based multi-target Docker deployment agent"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Kalpesh", email = "info@olgamack.com" }]

dependencies = [
    # API + runtime
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Agent framework
    "langgraph>=1.0.0",
    "langgraph-checkpoint-mongodb>=0.1.0",
    "langchain-openai>=0.2.0",

    # LLM client
    "openai>=1.50.0",

    # Integrations
    "PyGithub>=2.4.0",
    "docker>=7.1.0",
    "paramiko>=3.5.0",
    "python-telegram-bot>=21.0.0",

    # Database
    "motor>=3.6.0",
    "pymongo>=4.10.0",

    # Config + validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.2",

    # Observability
    "prometheus-client>=0.21.0",
    "structlog>=24.4.0",

    # Utilities
    "rapidfuzz>=3.10.0",
    "tabulate>=0.9.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "types-PyYAML>=6.0.0",
    "types-paramiko>=3.5.0",
    "types-tabulate>=0.9.0",
    "mongomock-motor>=0.0.34",
]

# ---------------------------------------------------------------------------
# setuptools — discover only the application packages, not tests/scripts/docs
# ---------------------------------------------------------------------------
[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "api*", "telegram_bot*", "tools*", "utils*", "config*"]
exclude = ["tests*", "scripts*", "docs*", "secrets*", "logs*", ".venv*"]

# ---------------------------------------------------------------------------
# ruff — lint + format
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = [".venv", "logs", "secrets"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "N",    # pep8-naming
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "ARG",  # flake8-unused-arguments
    "ASYNC",
]
ignore = [
    "E501",  # line-length handled by formatter
    "B008",  # FastAPI uses Depends() in defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# mypy — strict typing on first-party code
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
disallow_untyped_decorators = false  # langgraph/pydantic decorators

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-ra --strict-markers --strict-config"

[tool.coverage.run]
branch = true
source = ["agents", "api", "telegram_bot", "tools", "utils", "config"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
